Factor files_test into separate tests for each use case.
The motivation for separate tests is so that consumers of this library
can more easily patch out tests for various flags.  In particular, the
environment variable support may be troublesome for some environments
that want to enforce reproducible builds.

This change also consolidates on `_golden.json` naming for golden
files.  A beneficial side effect is that test failures will be easier
pin down.