Use the Go Jsonnet port instead of the C++ port (#113)
* Use the Go Jsonnet port instead of the C++ port

* Allow use of the C++ port via command-line flag

By default, Bazel will use the Go port of Jsonnet. To use the C++ port
of Jsonnet instead, invoke Bazel with the "--define jsonnet_port=cpp"
command-line flag. To select the Go port explicitly, invoke Bazel with
the "--define jsonnet_port=go" command-line flag.

* Change default Jsonnet port back from Go to C++

Retain use of C++ as the default port for one more release, with the
intention to change the default to the Go port and eventually remove
support for the C++ port.

* Use the latest version of the Jsonnet Go port

The selected commit is the latest along the "master" branch as of 24
August 2019.