Use Git to fetch Go Jsonnet port with submodules
Since the Go Jsonnet port uses Git submodules in its repository, we
can't use Bazel's http_archive function to fetch it, as that leaves us
with only the top-level Git repository content but without the
submodule content it requires. Return to using the git_repository
function, and initialize the submodules in the Git repository.